refactor: Split plugin.ts into multiple modules for better maintainability
